Add catch-type to rarities
