+++ even-more-fish-plugin/src/main/java/com/oheers/fish/fishing/CatchType.java
@@ -1,0 +1,7 @@
+package com.oheers.fish.fishing;
+
+public enum CatchType {
+    CATCH,
+    HUNT,
+    BOTH
+}
