+++ even-more-fish-plugin/src/main/java/com/oheers/fish/fishing/HuntingProcessor.java
@@ -101,8 +101,8 @@ public class HuntingProcessor extends Processor<EntityDeathEvent> {
 
     @Override
     public boolean canUseFish(@NotNull Fish fish) {
-        return fish.getCatchType().equals(Fish.CatchType.HUNT)
-            || fish.getCatchType().equals(Fish.CatchType.BOTH);
+        return fish.getCatchType().equals(CatchType.HUNT)
+            || fish.getCatchType().equals(CatchType.BOTH);
     }
 
 }
