+++ even-more-fish-plugin/src/main/resources/rarities/_example.yml
@@ -26,6 +26,10 @@ displayname: <yellow>Example Rarity
 # The permission required to fish this rarity.
 permission: emf.rarity.example
 
+# The catch type for this rarity. Defaults to CATCH if missing.
+# Current supported values: CATCH, HUNT, BOTH
+catch-type: CATCH
+
 # You can set custom requirements for this rarity to appear, all possible requirements are on the wiki: https://github.com/EvenMoreFish/EvenMoreFish/wiki/Requirements
 requirements:
   biome:
@@ -85,7 +89,7 @@ fish:
     # Lets you set a custom displayname for the fish, this replaces the {fish} in "You have caught a {rarity} {fish}" and the fish name in the item.
     displayname: <red>Example Fish
 
-    # The catch type for this fish. Current supported values: CATCH, HUNT, BOTH
+    # The catch type for this fish. This overrides the rarity's value.
     catch-type: CATCH
 
     # Should the fisherman be removed from the fish lore?
